@@ -2,18 +2,15 @@
   "name": "scs",
   "version": "0.0.1",
   "description": "Ship Control System - Empty Epsilon integration server",
-  "main": "dist/src/nodejs.js",
-  "browser": "dist/src/browser.js",
+  "main": "dist/src/server.js",
   "scripts": {
     "clean": "rimraf ./dist",
     "build": "tsc -d",
-    "build:watch": "tsc -d -w",
-    "bundle": "webpack --progress",
-    "pretest": "npm run clean && npm run build && npm run bundle",
-    "test": "npm run test:node && npm run test:browser",
-    "test:node": "mocha --reporter mocha-env-reporter dist/test/nodejs",
-    "test:browser": "karma start --single-run",
-    "start": "webpack-dev-server --progress"
+    "pretest": "npm run clean && npm run build",
+    "reset": "rimraf node_modules dist && yarn install && yarn test",
+    "test": "mocha --reporter mocha-env-reporter dist/test/index",
+    "start": "rimraf dist && yarn build && node ./main.js",
+    "debug": "rimraf dist && yarn build && concurrently --kill-others --success first \"yarn build -- -w\" \"node --inspect-brk ./main.js\""
   },
   "repository": {
     "type": "git",
@@ -30,38 +27,24 @@
   },
   "homepage": "https://github.com/CommaSword/scs#readme",
   "devDependencies": {
-    "@types/bluebird": "3.0.37",
-    "@types/bluebird-retry": "0.0.31",
     "@types/chai": "3.4.35",
     "@types/chai-as-promised": "0.0.29",
     "@types/lodash": "4.14.55",
     "@types/mocha": "2.2.39",
-    "@types/node": "7.0.5",
+    "@types/node": "8.0.33",
     "@types/socket.io-client": "1.4.29",
     "chai": "3.5.0",
     "chai-as-promised": "6.0.0",
     "chai-subset": "1.5.0",
-    "json-loader": "0.5.4",
-    "karma": "1.4.1",
-    "karma-chrome-launcher": "2.0.0",
-    "karma-env-reporter": "1.0.13",
-    "karma-firefox-launcher": "1.0.0",
-    "karma-mocha": "1.3.0",
+    "concurrently": "~3.5.0",
     "mocha": "3.2.0",
     "mocha-env-reporter": "2.0.4",
-    "mocha-loader": "1.1.0",
     "rimraf": "2.6.0",
     "sinon": "2.0.0-pre.2",
-    "source-map-support": "0.4.11",
-    "ts-loader": "2.0.0",
-    "typescript": "2.3.4",
-    "webpack": "1.14.0",
-    "webpack-dev-server": "1.16.2"
+    "typescript": "2.5.3"
   },
   "dependencies": {
     "axios": "0.15.3",
-    "bluebird": "3.4.7",
-    "bluebird-retry": "0.10.1",
     "eventemitter3": "2.0.2",
     "lodash": "4.17.4",
     "socket.io-client": "1.7.3",
