@@ -1,9 +1,9 @@
-import {Promise, ObjectDriver} from "../driver";
+import {ObjectDriver} from "../driver";
 import {SpaceShip} from "./space-ship";
 
 export class PlayerShip extends SpaceShip {
 
-    constructor(protected driver:ObjectDriver){
+    constructor(protected driver: ObjectDriver) {
         super();
     }
 
