@@ -1,4 +1,3 @@
-import {isUndefined} from "util";
 export type Unknown_MsgType = 'unknown';
 export type Noop_MsgType = 'noop';
 export type Hello_MsgType = 'hello';
@@ -7,38 +6,44 @@ export type State_MsgType = 'state';
 export type MsgType = State_MsgType | Hello_MsgType | Noop_MsgType | Unknown_MsgType;
 
 export const delimitter = ';';
+
 export interface Msg<T extends MsgType> {
     type: T;
 }
+
 export interface IncomingMsg<T extends MsgType> extends Msg<T> {
-    [k:string]:any;
+    [k: string]: any;
 }
 
-export interface NoopMsg extends IncomingMsg<Noop_MsgType>{
-    id:string;
+export interface NoopMsg extends IncomingMsg<Noop_MsgType> {
+    id: string;
 }
-export function isNoopMsg(msg:Msg<any>):msg is NoopMsg{
+
+export function isNoopMsg(msg: Msg<any>): msg is NoopMsg {
     return msg.type === 'noop';
 }
 
-export interface HelloMsg extends IncomingMsg<Hello_MsgType>{
-    id:string;
-    state:any;
+export interface HelloMsg extends IncomingMsg<Hello_MsgType> {
+    id: string;
+    state: any;
 }
-export function isHelloMsg(msg:Msg<any>):msg is HelloMsg{
+
+export function isHelloMsg(msg: Msg<any>): msg is HelloMsg {
     return msg.type === 'hello';
 }
 
-export function validateHelloMsg(msg:Msg<any>):msg is HelloMsg{
+export function validateHelloMsg(msg: Msg<any>): msg is HelloMsg {
     return isHelloMsg(msg) && typeof msg.id === 'string' && msg.state !== undefined;
 }
 
-export interface StateMsg extends IncomingMsg<State_MsgType>{
-    state:any;
+export interface StateMsg extends IncomingMsg<State_MsgType> {
+    state: any;
 }
-export function isStateMsg(msg:Msg<any>):msg is StateMsg{
+
+export function isStateMsg(msg: Msg<any>): msg is StateMsg {
     return msg.type === 'state';
 }
-export function validateStateMsg(msg:Msg<any>):msg is StateMsg{
+
+export function validateStateMsg(msg: Msg<any>): msg is StateMsg {
     return isStateMsg(msg) && msg.state !== undefined;
 }
