@@ -1,3 +1,3 @@
-export const runServer =  "killall EmptyEpsilon; /Users/amira/ee/emptyepsilon/EmptyEpsilon/EmptyEpsilon.app/Contents/MacOS/EmptyEpsilon headless=amir.lua headless_name=Federation httpserver=8080";
-export const killServer = "killall EmptyEpsilon";
-export const serverAddress = "http://localhost:8080";
+export const runServer =  "/Applications/EmptyEpsilon.app/Contents/MacOS/EmptyEpsilon headless=scenario_00_basic.lua headless_name=Federation httpserver=8081";
+export const killServer = "killall EmptyEpsilon & echo EE server killed";
+export const serverAddress = "http://localhost:8081";
